requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
requests
Django==3.0.5
pytest
pytest-pythonpath==0.7.3
pytest-django
python-dotenv==0.19.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-filter
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1